+++ lib/features/post_details/logic/post_details_logic/job_details_state.freezed.dart
