+++ lib/features/post_details/ui/job_details/widgets/job_details/coin_icon_and_salary.dart
