+++ lib/features/post_details/ui/job_details/widgets/job_details/job_name_location_type_post_date.dart
@@ -4,9 +4,8 @@ import 'package:carrerk/core/theming/styles.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-
 import '../../../../../../../../core/helpers/date_formatter.dart';
-import '../../../data/models/job_details_response_body.dart';
+import '../../../../data/models/post_details_models/job_details_response_body.dart';
 
 class JobNameLocationTypePostDate extends StatelessWidget {
   final JobDetailsResponseBody data;
