+++ lib/features/post_details/ui/job_details/widgets/job_details_shimmer/coin_icon_and_salary_shimmer.dart
