+++ lib/features/post_details/ui/job_details/widgets/job_details_shimmer/job_description_shimmer.dart
