+++ lib/features/post_details/ui/job_details/widgets/job_details_shimmer/job_name_location_type_post_date_shimmer.dart
